requirements.txt
fastapi==0.63.0
motor==2.2.0
pydantic[email]
python-decouple==3.4
uvicorn==0.13.4
pytest==6.2.2
requests
pyjwt
passlib
bcrypt